% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrawl_plot.R
\name{scrawl_plot}
\alias{scrawl_plot}
\title{Construct a scrawl plot from the scrawl object}
\usage{
scrawl_plot(data, palette = "grayC")
}
\arguments{
\item{data}{A tibble generated by scrawl_build()}

\item{palette}{String specifying a scico palette}
}
\value{
A ggplot object
}
\description{
Construct a scrawl plot from the scrawl object
}
