% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrawl_save.R
\name{scrawl_save}
\alias{scrawl_save}
\title{Save a scrawl plot to file}
\usage{
scrawl_save(plot, filename, pixels = 5000)
}
\arguments{
\item{plot}{Object to save}

\item{filename}{Path to file}

\item{pixels}{Resolution of image in pixels}
}
\value{
Invisibly returns NULL
}
\description{
Save a scrawl plot to file
}
\details{
This is just a thin wrapper around ggsave()
}
